refactor(convertors): split conversion helpers into internal/convtypes
Move the schema-independent primitive helpers (StringPointer,
TimePointerFromString, …) out of the conv package into a new leaf package
internal/convtypes. The generated Layer 3 convertor now imports it and qualifies
helper calls (convtypes.StringPointer, …); conv keeps only the schema-coupled
generated code. Mirrors the internal/flex layering used by larger providers.

Regenerated via the new `skaff layer3 --convtypes-import` flag, which qualifies
emitted helper calls against the convtypes package.

The unused generic Expander/Flattener contract interfaces are dropped for now.

go build ./..., go vet ./internal/..., and golangci-lint all pass.
